+++ src/agixt/__init__.py
@@ -6,4 +6,3 @@ from .Config.Agent import Agent
 from .Config import Config
 from .Embedding import Embedding
 from .CustomPrompt import CustomPrompt
-from .Learning import Learning
